fix(distributed): thread explicit Time through snapshot and sync paths
create_snapshot hardcoded Time::ZERO for the snapshot timestamp,
making all snapshots appear to have been created at epoch zero
regardless of when they were actually taken.  Similarly,
upgrade_to_distributed did not accept a time parameter, leaving
last_sync_time stale after the upgrade.

Fix: both create_snapshot and upgrade_to_distributed now take an
explicit now: Time argument that is threaded through to the snapshot
and sync state.  All call sites (bridge.rs, federation.rs,
e2e_distributed.rs, fabric_e2e.rs) updated to pass real timestamps.

Tests updated to pass specific Time values and assert that snapshot
timestamps and last_sync_time reflect the provided time rather than
defaulting to zero.

+++ src/distributed/bridge.rs
@@ -1550,11 +1550,11 @@ mod tests {
         let mut bridge = create_local_bridge();
 
         let mut prev_seq = 0;
-        for i in 0u64..20 {
+        for i in 0u32..20 {
             // Interleave task add/remove with snapshots.
-            let tid = TaskId::new_for_test(i as u32, 0);
+            let tid = TaskId::new_for_test(i, 0);
             bridge.add_task(tid).unwrap();
-            let snap = bridge.create_snapshot(Time::from_secs(i + 1));
+            let snap = bridge.create_snapshot(Time::from_secs(u64::from(i) + 1));
             assert!(
                 snap.sequence > prev_seq,
                 "sequence must be monotonically increasing"
